// backend/src/enums/platform-permission.enum.ts
export enum PlatformPermissionEnum {
    LISTING_CREATE = "PLATFORM:LISTING:CREATE",
    LISTING_EDIT = "PLATFORM:LISTING:EDIT",
    LISTING_DELETE = "PLATFORM:LISTING:DELETE",
    LISTING_ACTIVATE = "PLATFORM:LISTING:ACTIVATE",
    LISTING_DEACTIVATE = "PLATFORM:LISTING:DEACTIVATE",
    LISTING_GET_STATS = "PLATFORM:LISTING:GET_STATS",

    USER_GET_ALL = "PLATFORM:USER:GET_ALL",
    USER_EDIT_BY_ADMIN = "PLATFORM:USER:EDIT_BY_ADMIN",
    USER_ACTIVATE = "PLATFORM:USER_ACTIVATE",
    USER_DEACTIVATE = "PLATFORM:USER_DEACTIVATE",
    USER_DELETE = "PLATFORM:USER:DELETE",

    ME_GET = "PLATFORM:ME:GET",
    ME_UPDATE = "PLATFORM:ME:UPDATE",
    ME_TOP_UP = "PLATFORM:ME:TOP_UP",
    ME_BECOME_SELLER = "PLATFORM:ME:BECOME_SELLER",
    ME_UPGRADE_PLAN = "PLATFORM:ME:UPGRADE_PLAN",
    ME_DEACTIVATE = "PLATFORM:ME:DEACTIVATE",
    ME_LISTING_DEACTIVATE = "PLATFORM:ME:LISTING:DEACTIVATE",
}